Remove initial nonce increment on create
create account doesn't increment nonce automatically.
The nonce is incremented if it is created through a UserOp

Side-effect:
if account is created NOT through AA, then first UserOp will cost 20000
gas more.
(if created as a userop, then this extra cost is part of the "creation"
and not on a separate transaction.