Disallow paymaster for staked account
Staked account could have more than one userOp in the same bundler, so it could potentially grieve the paymaster if two userOps (from the same sender account)
are trying to empty its balance.

"
A sender is exempt from this rule and may have multiple UserOperations in the pool and in a batch if it is staked (see reputation, throttling and banning section below), but this exception is of limited use to normal accounts.
"